Add files via upload
Category : classification
Description : Main purpose: This algorithm is used to predict the probability of a sample belonging to a particular class. Its output is a value between 0 and 1, which can be interpreted as a probability.
Sigmoid function: Logistic Regression uses a function called the sigmoid function to map linear inputs to a probability output between 0 and 1.
Binary classification: By default, Logistic Regression is used for binary classification tasks, meaning it determines whether a sample belongs to class 0 or class 1.